+++ context-propagation/src/test/java/io/micrometer/context/ScopedValueSnapshotTests.java
@@ -19,10 +19,10 @@ import java.util.Collections;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicReference;
 
-import io.micrometer.context.scopedvalue.Scope;
-import io.micrometer.context.scopedvalue.ScopedValue;
-import io.micrometer.context.scopedvalue.ScopeHolder;
-import io.micrometer.context.scopedvalue.ScopedValueThreadLocalAccessor;
+import io.micrometer.scopedvalue.Scope;
+import io.micrometer.scopedvalue.ScopedValue;
+import io.micrometer.scopedvalue.ScopeHolder;
+import io.micrometer.scopedvalue.ScopedValueThreadLocalAccessor;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
