+++ context-propagation/src/test/java/io/micrometer/scopedvalue/SimpleScopedValue.java
@@ -13,21 +13,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micrometer.context.scopedvalue;
+package io.micrometer.scopedvalue;
 
 /**
- * Implementation of {@link ScopedValue} which is used when the actual value in scope
- * should be missing, but the scope hierarchy needs to be maintained.
+ * Implementation of {@link ScopedValue} for which {@link Scope} maintains the hierarchy
+ * between parent scope and an opened scope for this value.
  */
-class NullScopedValue extends SimpleScopedValue {
+class SimpleScopedValue implements ScopedValue {
 
-    NullScopedValue() {
-        super("null");
+    private final String value;
+
+    SimpleScopedValue(String value) {
+        this.value = value;
     }
 
     @Override
     public String get() {
-        return null;
+        return value;
     }
 
 }
